+++ tools/src/librtools/RtimerFd.cpp
@@ -1,9 +1,10 @@
-// $Id: RtimerFd.cpp 1185 2019-07-12 17:29:12Z mueller $
+// $Id: RtimerFd.cpp 1376 2023-02-20 15:05:03Z mueller $
 // SPDX-License-Identifier: GPL-3.0-or-later
-// Copyright 2013-2019 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
+// Copyright 2013-2023 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 // 
 // Revision History: 
 // Date         Rev Version  Comment
+// 2023-02-20  1376   1.1.1  BUGFIX: SetRelative(): correct is-positive check
 // 2019-06-08  1161   1.1    derive from Rfd, inherit IsOpen,Close,Fd
 // 2017-02-18   852   1.0    Initial version
 // 2013-01-11   473   0.1    First draft
@@ -67,7 +68,7 @@ void RtimerFd::SetRelative(const Rtime& dt)
   if (!IsOpen())
     throw Rexception(fCnam+"SetRelative()", "bad state: not open");
 
-  if (dt.Sec() <= 0 || dt.NSec() <= 0)
+  if (!dt.IsPositive())
     throw Rexception(fCnam+"SetRelative()", "bad value: dt zero or negative ");
 
   struct itimerspec itspec;